requirements.txt
alabaster==0.7.9
argh==0.26.2
Babel==2.3.4
backports-abc==0.4
certifi==2016.8.2
coverage==4.0.3
cycler==0.10.0
decorator==4.0.10
Django==1.10
django-extensions==1.7.6
django-filter==0.12.0
django-mathfilters==0.4.0
djangorestframework==3.3.2
docutils==0.12
eventlet==0.19.0
geopy==1.11.0
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
haversine==0.4.5
imagesize==0.7.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
livereload==2.4.1
Markdown==2.6.5
MarkupSafe==0.23
matplotlib==1.5.1
numpy==1.11.0
pandas==0.19.2
pathtools==0.1.2
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.1.1
port-for==0.3.1
prompt-toolkit==1.0.5
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
scikit-learn==0.17.1
scipy==0.17.1
Shapely==1.5.17.post1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.4.5
sphinx-autobuild==0.6.0
sphinx-rtd-theme==0.1.9
tornado==4.4.1
traitlets==4.2.2
watchdog==0.8.3
wcwidth==0.1.7
